Fix a bug with image loading for 'norewind' calls.
This CL fixes an issue related to invoking `justifiedGallery('norewind')` on a collection of items where the new items already have their images loaded. This issue only happens when using waitThumbnailsLoad = false.

To fix it I've pulled out the (now) common code used to listen for load events on an image into a new onImageEvent method that both call site now use to register interesting in the image load.