Added option to justify images in last row to average height
Imagine a small portrait screen displaying relative big images, may less
then tree per row. Most of them are required to scale up to fit the area
of the gallery. If the last row is flushed e.g. in nojustify mode (=
default) it gets significant smaller then the others. This looks
annoying, especially on this kind of device. Using justify mode would
fix it in this case, but not acceptable for big screens.

To comply both requirements there are new options (left-average, center-
average and right-average) which cause the last row is resized to the
average row size.