Fix width/height property access
This makes it possible to eliminate one set of roundstrips (for default
sized thumbnails) by specifying the height/width in the <img> tag and
setting src to a placeholder image (e.g. a 1x1 data-uri gif), which lets
JustifiedGallery request the right sized thumbnails directly. Without
this commit, the images ended up with the placeholders aspect ratio
(i.e., square).